# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SendPrint API — receives print jobs from Promesse and forwards ZPL to a Zebra label printer
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface PrintJobInput {
  /** Raw ZPL commands to send to the printer */
  zpl: string;
  /** Optional source identifier (e.g. "Promesse") */
  source?: string;
}

export interface PrintJobResult {
  success: boolean;
  jobId: number;
  message?: string;
  bytesSent?: number;
}

export type PrintJobStatus = typeof PrintJobStatus[keyof typeof PrintJobStatus];


export const PrintJobStatus = {
  success: 'success',
  failed: 'failed',
} as const;

export interface PrintJob {
  id: number;
  createdAt: string;
  status: PrintJobStatus;
  /** @nullable */
  source?: string | null;
  /** @nullable */
  zplPreview?: string | null;
  bytesSent: number;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  printerIp?: string | null;
}

export interface PrinterStatus {
  online: boolean;
  printerIp: string;
  printerPort: number;
  checkedAt: string;
  /** @nullable */
  message?: string | null;
}

export interface Config {
  printerIp: string;
  /** TCP port on the Zebra printer (default 9100) */
  printerPort: number;
  /** HTTP port the SendPrint bridge listens on (takes effect after restart) */
  serverPort: number;
  apiKey: string;
  updatedAt: string;
}

export interface ConfigUpdate {
  /** @minLength 1 */
  printerIp?: string;
  /**
     * @minimum 1
     * @maximum 65535
     */
  printerPort?: number;
  /**
     * @minimum 1
     * @maximum 65535
     */
  serverPort?: number;
  /** @minLength 8 */
  apiKey?: string;
}

export interface LogStats {
  totalJobs: number;
  successCount: number;
  failedCount: number;
  last24hCount: number;
  /** @nullable */
  lastJobAt: string | null;
}

export interface ClearResult {
  deleted: number;
}

export interface AuthStatus {
  needsSetup: boolean;
  authenticated: boolean;
  /** @nullable */
  expiresAt?: string | null;
}

export interface AuthResult {
  success: boolean;
}

export interface AuthLoginBody {
  /** @minLength 1 */
  password: string;
}

export interface AuthSetupBody {
  /** @minLength 8 */
  password: string;
}

export interface AuthChangePasswordBody {
  /** @minLength 1 */
  currentPassword: string;
  /** @minLength 8 */
  newPassword: string;
}

export type GetLogsParams = {
/**
 * @minimum 1
 * @maximum 500
 */
limit?: number;
};