# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SendPrint API — receives print jobs from Promesse and forwards ZPL to a Zebra label printer
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/setup:
    post:
      operationId: setupAdmin
      tags: [auth]
      summary: Create the initial admin password (first-run only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthSetupBody"
      responses:
        "200":
          description: Admin created and signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "409":
          description: Admin already configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in as admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthLoginBody"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change the admin password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthChangePasswordBody"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Not authenticated or current password incorrect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    AuthResult:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
    AuthLoginBody:
      type: object
      required:
        - password
      properties:
        password:
          type: string
          minLength: 1
    AuthSetupBody:
      type: object
      required:
        - password
      properties:
        password:
          type: string
          minLength: 8
    AuthChangePasswordBody:
      type: object
      required:
        - currentPassword
        - newPassword
      properties:
        currentPassword:
          type: string
          minLength: 1
        newPassword:
          type: string
          minLength: 8