# lib/api-zod/src/generated/types/authChangePasswordBody.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SendPrint API — receives print jobs from Promesse and forwards ZPL to a Zebra label printer
 * OpenAPI spec version: 0.1.0
 */

export interface AuthChangePasswordBody {
  /** @minLength 1 */
  currentPassword: string;
  /** @minLength 8 */
  newPassword: string;
}